Rearrange/repackage controls.
