+++ src/commonMain/kotlin/baaahs/control/VisualizerControl.kt
@@ -1,0 +1,82 @@
+package baaahs.control
+
+import baaahs.Gadget
+import baaahs.ShowPlayer
+import baaahs.app.ui.ControlEditIntent
+import baaahs.app.ui.EditIntent
+import baaahs.app.ui.EditorPanel
+import baaahs.app.ui.editor.EditableManager
+import baaahs.app.ui.editor.GenericPropertiesEditorPanel
+import baaahs.app.ui.editor.VisualizerPropsEditor
+import baaahs.show.Control
+import baaahs.show.live.ControlProps
+import baaahs.show.live.OpenContext
+import baaahs.show.live.OpenControl
+import baaahs.show.live.controlViews
+import baaahs.show.mutable.MutableControl
+import baaahs.show.mutable.MutableShow
+import baaahs.show.mutable.ShowBuilder
+import baaahs.ui.Renderer
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+@SerialName("baaahs.Core:Visualizer")
+data class VisualizerControl(
+    val surfaceDisplayMode: SurfaceDisplayMode = SurfaceDisplayMode.Continuous,
+    val rotate: Boolean = false
+): Control {
+    enum class SurfaceDisplayMode {
+        Pixels,
+        Continuous
+    }
+
+    override val title: String get() = "Visualizer"
+
+    override fun createMutable(mutableShow: MutableShow): MutableVisualizerControl {
+        return MutableVisualizerControl(surfaceDisplayMode, rotate)
+    }
+
+    override fun open(id: String, openContext: OpenContext, showPlayer: ShowPlayer): OpenControl {
+        return OpenVisualizerControl(id, this)
+    }
+}
+
+data class MutableVisualizerControl(
+    var surfaceDisplayMode: VisualizerControl.SurfaceDisplayMode = VisualizerControl.SurfaceDisplayMode.Continuous,
+    var rotate: Boolean = false
+) : MutableControl {
+    override val title: String get() = "Visualizer"
+
+    override var asBuiltId: String? = null
+
+    override fun getEditorPanels(editableManager: EditableManager): List<EditorPanel> {
+        return listOf(
+            GenericPropertiesEditorPanel(
+                editableManager,
+                VisualizerPropsEditor(this)
+            )
+        )
+    }
+
+    override fun build(showBuilder: ShowBuilder): Control {
+        return VisualizerControl(surfaceDisplayMode, rotate)
+    }
+}
+class OpenVisualizerControl(
+    override val id: String,
+    private val visualizerControl: VisualizerControl
+) : OpenControl {
+    override val gadget: Gadget? get() = null
+
+    override fun toNewMutable(mutableShow: MutableShow): MutableControl =
+        visualizerControl.createMutable(mutableShow)
+
+    override fun getRenderer(controlProps: ControlProps): Renderer =
+        controlViews.forVisualizer(this, controlProps)
+
+    override fun getEditIntent(): EditIntent =
+        ControlEditIntent(id)
+
+    val rotate get() = visualizerControl.rotate
+}
