+++ src/commonTest/kotlin/baaahs/show/live/TestUtil.kt
@@ -2,6 +2,8 @@ package baaahs.show.live
 
 import baaahs.ShowState
 import baaahs.app.ui.editor.PortLinkOption
+import baaahs.control.OpenButtonControl
+import baaahs.control.OpenButtonGroupControl
 import baaahs.gl.Toolchain
 import baaahs.gl.autoWire
 import baaahs.gl.shader.type.PaintShader
